requirements.txt
requests
python-dotenv
pycrypto